examples/keywords_extractor_agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.0
pydantic>=2.0.0

